requirements.txt
streamlit
python-dotenv
langchain-openai
langchain-core
pydantic 